HUGE STRING FIX (see full commit for more info)
Strings have always been handled badly in Sporth.

Previously, strings were copied to/from the stack into char arrays
with a fixed length of 256.

While it fixed some problems in the early days, it
made things quite messy:

- Strings were limited to only 256 characters
- Strings had to be freed from the ugen (very sketchy)
- Wasteful memory: Each stack entry had it's own 256 character array.

Strings in Sporth have been re-factored in such a way that
the stack holds references to strings stored in the plumbing
instead of copying strings every time. Obvious in retrospect now,
but not during the initial implementation.

What this means is that I will need to go in and modify all
the ugens that use strings so they don't call free. I've
tested things with gen_vals and things seem to work fine.
Valgrind passes. GDB shows that strings are being preserved.