Variable tweaks for optimization.
I have made some changes from the initial idea in order to make
variables faster.

Variables are now stored as pure SPFLOATs instead of 1-sized
soundpipe ftables, reducing the amount of memory needed.

The downside of this is that variables are not compatible with ftable
ugens like tget, tset, tbldur, tblsize, tblrec, or tabread. But this
is a small price to pay, since you shouldn't actually be using these
there.